Move block locators, add test
