+++ node/messages/Cargo.toml
@@ -25,6 +25,9 @@ version = "1.0"
 [dependencies.bytes]
 version = "1"
 
+[dependencies.indexmap]
+version = "1"
+
 [dependencies.kadmium]
 version = "0.6.0"
 features = ["codec", "sync"]
