+++ node/messages/src/helpers/mod.rs
@@ -14,6 +14,9 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkOS library. If not, see <https://www.gnu.org/licenses/>.
 
+mod block_locators;
+pub use block_locators::*;
+
 mod codec;
 pub use codec::MessageCodec;
